+++ src/main/java/racingcar/view/component/Component.java
@@ -1,4 +1,4 @@
-package racingcar.component;
+package racingcar.view.component;
 
 public interface Component {
     String render();
