+++ src/main/java/racingcar/view/component/Composable.java
@@ -1,4 +1,4 @@
-package racingcar.component;
+package racingcar.view.component;
 
 public interface Composable {
     Component compose();
