+++ src/main/java/racingcar/view/component/output/RoundResultComponent.java
@@ -1,12 +1,12 @@
-package racingcar.component.output;
+package racingcar.view.component.output;
 
 import static strings.Strings.DASH;
 import static strings.Strings.LINE_BREAK;
 
 import java.util.stream.Collectors;
-import racingcar.component.Component;
 import racingcar.domain.car.dto.output.CarRacingDto;
 import racingcar.domain.car.dto.output.CarsRacingDto;
+import racingcar.view.component.Component;
 import strings.Strings;
 
 public final class RoundResultComponent implements Component {
